bugfixes with sequential file loads and preppting hrm/cad/pwr data arrays
All gpx and fit files are now checked for power, cadence, and HR, and that data is logged in indexed arrays (index aligns with the i array, which is the same as x, y, elev, etc data already in use).
Also did some minor cleanup, and worked to resolve issues with loading files while prior animation was running.  Got this behavior improved dramatically, BUT realized it was silly - the animation is 2 seconds.  So I just hid there button while animation is running, and re-expose it when finish.  That way user can't force a file to load while prior is still running.